ingest_service/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
pydantic-settings>=2.2.0
aiokafka>=0.10.0
grpcio>=1.62.0
grpcio-tools>=1.62.0
protobuf>=4.25.0
httpx>=0.27.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
grpcio-testing>=1.62.0
